feat(web): give integrations their own page and move Daemons to the rail's foot
The org's messaging bots and code hosts were buried at the bottom of Settings,
a page that no-auth deployments cannot even reach (the shell bounces /settings
and /profile when auth is off) — so a self-hosted console had no way to see its
own bots or GitHub installations. They now have a top-level Integrations page,
which is reachable in both modes.

Rail: the nav is two groups behind a rule instead of one flat list. Daemons
moves to the foot on its own, because it is infrastructure you visit when
something is wrong, not a surface you work in; Integrations joins the first
group. The mobile More sheet mirrors the order.

Integrations page: the tabbed bots card and the GitHub installations card are
MOVED out of SettingsView, not copied — the same org infrastructure rendered on
two pages would be two sources of truth. Settings keeps organization, members
and the org-wide policies. The bots card gains per-platform counts and a
"Show in use" filter (counts stay unfiltered, so reading across the strip still
answers "where does this org have bots"); the GitHub card gains a header row.
The agent page's bot chip deep-links to /integrations?bot=…, and `integrations`
joins the CP's reserved org slugs.

Add integration: one dialog, two arms. Opened from an agent it is unchanged —
the agent is fixed and no picker renders. Opened from this page there is no
implied agent, so ModalProvider renders AddIntegrationForOrgModal, which owns
the selection and passes `agentChoices`; that prop is what makes the Agent field
appear. The dialog is remounted per agent so nothing agent-scoped (watched
repos, authorizations, the picked bot) survives a switch, and the picker offers
only agents the viewer can edit, since creating an integration writes the spec.

Verified light/dark, desktop and 375px, and both dialog arms in the browser.
