+++ packages/web/src/app/(app)/[slug]/integrations/page.tsx
@@ -1,0 +1,8 @@
+import type { Metadata } from 'next'
+import IntegrationsView from '@/components/console/views/IntegrationsView'
+
+export const metadata: Metadata = { title: 'Integrations · AgentConnect' }
+
+export default function Page() {
+  return <IntegrationsView />
+}
