+++ packages/web/src/components/console/views/AgentDetailView.tsx
@@ -499,7 +499,7 @@ export default function AgentDetailView() {
   const tabCls = (t: DetailTab) => (tab === t ? 'tab on' : 'tab')
   const tabHref = (t: DetailTab) => orgPath(t === 'integrations' ? `/agents/${da.id}` : `/agents/${da.id}?tab=${t}`)
   const botSettingsHref = (botId?: string) =>
-    orgPath(botId ? `/settings?bot=${encodeURIComponent(botId)}` : '/settings')
+    orgPath(botId ? `/integrations?bot=${encodeURIComponent(botId)}` : '/integrations')
 
   // ── Single responsive tree. Base classes are the mobile (≤768px) push-detail
   // body (the Shell provides the top push bar there); `desktop:` variants restore
